backend/requirements.txt
asgiref==3.9.1
Django==5.0.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==22.0.0
packaging==25.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
